requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
Django==2.2
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
importlib-metadata==4.6.4
Markdown==3.3.4
numpy==1.21.2
Pillow==8.3.1
psycopg2==2.8.6
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
